requirements.txt
Flask
Flask-SQLAlchemy
Flask-Markdown
requests
PyYAML
Faker
pygments
lxml